move files properly
