+++ pkg/execution/queue/option.go
@@ -734,46 +734,6 @@ type ShadowContinuation struct {
 	Count      uint
 }
 
-type DispatchedItem interface {
-	Done() <-chan DispatchedItemResult
-}
-
-type DispatchedItemResult struct {
-	ScheduledImmediateJob bool
-	Err                   error
-}
-
-type dispatchedItemHandle struct {
-	once sync.Once
-	done chan DispatchedItemResult
-}
-
-func newDispatchedItemHandle() *dispatchedItemHandle {
-	return &dispatchedItemHandle{
-		// Buffered so worker completion never blocks if no scanner is waiting.
-		done: make(chan DispatchedItemResult, 1),
-	}
-}
-
-// NewCompletedDispatchedItem returns a dispatched item that has already completed.
-// This is useful for scanner implementations or tests that synchronously process dispatch.
-func NewCompletedDispatchedItem(result DispatchedItemResult) DispatchedItem {
-	handle := newDispatchedItemHandle()
-	handle.complete(result)
-	return handle
-}
-
-func (h *dispatchedItemHandle) Done() <-chan DispatchedItemResult {
-	return h.done
-}
-
-func (h *dispatchedItemHandle) complete(result DispatchedItemResult) {
-	h.once.Do(func() {
-		h.done <- result
-		close(h.done)
-	})
-}
-
 // ProcessItem references the queue item and scanner-provided metadata to be processed by a worker.
 type ProcessItem struct {
 	I QueueItem
